update references from tests to pyloci
