+++ pyloci/sparql/query_loci_thing.py
@@ -4,42 +4,7 @@ import os
 from dotenv import load_dotenv, find_dotenv
 load_dotenv(find_dotenv())
 
-
-def query_sfWithin(loci_thing, sparql_endpoint, auth=None):
-    '''
-        auth = expects a dict with 'user' and 'password' as keys with values,
-              e.g. { 'user': 'username', 'password': 'passwordhere' }
-    '''
-    sparql = SPARQLWrapper(sparql_endpoint)
-
-    if auth !=  None:
-        sparql.setCredentials(user=auth['user'], passwd=auth['password'])
-
-    query = '''
-        PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
-        PREFIX geo: <http://www.opengis.net/ont/geosparql#>
-        PREFIX dct: <http://purl.org/dc/terms/>
-        select ?st ?link ?par
-        where {{
-            ?st dct:isPartOf ?lk ;
-                rdf:subject {loci_subject} ;
-                rdf:predicate geo:sfWithin ;
-                rdf:object ?par .
-        }}
-    '''.format(loci_subject=loci_thing)
-    #print(query)
-    sparql.setQuery(query)
-    sparql.setReturnFormat(JSON)
-    results = sparql.query().convert()
-    #print(json.dumps(results, indent=4, sort_keys=True))
-
-    matching_list = []
-    for r in results["results"]["bindings"]:
-        if 'par' in r:
-            matching_list.append(r['par']['value'])
-
-    #do something    
-    return matching_list
+from . import util
 
     
 
@@ -74,7 +39,7 @@ loci_thing_list = [
 print(auth)
 matches = []
 for curr in loci_thing_list:
-    res = query_sfWithin(curr, SPARQL_ENDPOINT, auth=auth)
+    res = util.query_sfWithin(curr, SPARQL_ENDPOINT, auth=auth)
     matches.append({"id" : curr, "matches": res})
 
 print(json.dumps(matches, indent=4, sort_keys=True))
