+++ pyloci/sparql/query_loci_withins.py
@@ -5,80 +5,7 @@ import csv
 from dotenv import load_dotenv, find_dotenv
 load_dotenv(find_dotenv())
 
-
-def query_sfWithin(locationUri, sparql_endpoint, auth=None):
-    '''
-        auth = expects a dict with 'user' and 'password' as keys with values,
-              e.g. { 'user': 'username', 'password': 'passwordhere' }
-    '''
-    sparql = SPARQLWrapper(sparql_endpoint)
-
-    if auth !=  None:
-        sparql.setCredentials(user=auth['user'], passwd=auth['password'])
-
-    cc = '?cc'
-    mb = '?mb'
-    if 'asgs2016/meshblock' in locationUri:
-        mb = locationUri
-        filterStmt = 'FILTER (?mb = ' + mb + ')'
-    elif 'geofabric/contractedcatchment' in locationUri:
-        cc = locationUri
-        filterStmt = 'FILTER (?cc = ' + cc + ')'
-
-    query = '''
-        PREFIX void: <http://rdfs.org/ns/void#>
-        PREFIX loci: <http://linked.data.gov.au/def/loci#>
-        PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
-        PREFIX geo: <http://www.opengis.net/ont/geosparql#>
-        PREFIX dct: <http://purl.org/dc/terms/>
-        prefix dbp: <http://dbpedia.org/property/>
-        PREFIX nv: <http://qudt.org/schema/qudt#numericValue>
-        PREFIX qu: <http://qudt.org/schema/qudt#unit>
-        PREFIX asgs: <http://linked.data.gov.au/def/asgs#>
-        PREFIX data: <http://linked.data.gov.au/def/datatype/> 
-        PREFIX geox: <http://linked.data.gov.au/def/geox#>
-        PREFIX g: <http://linked.data.gov.au/dataset/gnaf/address/>
-        SELECT DISTINCT ?mb ?cc ?mbArea ?ccArea
-        WHERE {{
-            ?s dct:isPartOf <http://linked.data.gov.au/dataset/mb16cc> ;
-            rdf:subject ?mb ;
-            rdf:predicate geo:sfWithin ;
-            rdf:object ?cc .
-            ?mb a asgs:MeshBlock
-            OPTIONAL {{
-                ?mb geox:hasAreaM2 [ data:value ?mbArea ] .
-            }}
-            GRAPH <http://linked.data.gov.au/dataset/mb16cc> {{
-                OPTIONAL {{
-                    ?cc geox:hasAreaM2 [ 
-                        data:value ?ccArea ;
-                    ] .                
-                }}
-            }}
-            {filterStmt}
-        }}
-    '''.format(filterStmt=filterStmt)
-    #print(query)
-    sparql.setQuery(query)
-    sparql.setReturnFormat(JSON)
-    results = sparql.query().convert()
-    #print(json.dumps(results, indent=4, sort_keys=True))
-
-    withins_list = []
-    for res in results['results']['bindings']:
-        #print(json.dumps(res, indent=4, sort_keys=True))
-        #print(res['mb']['value'] + "," + res['cc']['value'])
-        #encode <from URI, URI matching withins for from URI>
-        withins_list.append( {
-                    'mb': res['mb']['value'], 
-                    'cc': res['cc']['value'], 
-                    'mbArea': res['mbArea']['value'], 
-                    'ccArea' : res['ccArea']['value']
-                }
-            )
-
-    #do something    
-    return withins_list
+from . import util
 
 def parse_input_file(csvfile):   
     test_list = []
@@ -127,7 +54,7 @@ for (testcase,uristr) in loci_cc_list:
     uri = "<" + uristr + ">"
     #print(testcase)
     #print(uristr)
-    withins_list = query_sfWithin(uri, SPARQL_ENDPOINT, auth=auth)
+    withins_list = util.query_sfWithin_mb_or_cc(uri, SPARQL_ENDPOINT, auth=auth)
     if testcase in test_case_withins_list:
         test_case_withins_list[testcase][uristr] = withins_list
     else:
