+++ pyloci/sparql/query_repo.py
@@ -4,6 +4,8 @@ import os
 from dotenv import load_dotenv, find_dotenv
 load_dotenv(find_dotenv())
 
+from . import util
+
 
 def query(sparql_endpoint):
     sparql = SPARQLWrapper(sparql_endpoint)
@@ -20,34 +22,7 @@ def query(sparql_endpoint):
 
     print(results)
     
-def query_type(loci_type, sparql_endpoint, auth=None):
-    '''
-        auth = expects a dict with 'user' and 'password' as keys with values,
-              e.g. { 'user': 'username', 'password': 'passwordhere' }
-    '''
-    sparql = SPARQLWrapper(sparql_endpoint)
 
-    if auth !=  None:
-        sparql.setCredentials(user=auth['user'], passwd=auth['password'])
-
-    query = '''
-        PREFIX geo: <http://www.opengis.net/ont/geosparql#>
-        PREFIX loci: <http://linked.data.gov.au/def/loci#>
-        PREFIX hyf: <https://www.opengis.net/def/appschema/hy_features/hyf/>
-        PREFIX dcat: <http://www.w3.org/ns/dcat#>
-        select (count(distinct ?x) as ?count)
-        where {{
-            ?x a {definedtype}
-        }}
-    '''.format(definedtype=loci_type)
-    #print(query)
-    sparql.setQuery(query)
-    sparql.setReturnFormat(JSON)
-    results = sparql.query().convert()
-    
-    count  = results["results"]["bindings"][0]['count']['value']
-    print(loci_type + ", " + count)
-    return count
     
 
 
@@ -89,7 +64,7 @@ loci_types = [
 print(auth)
 count_types = []
 for curr in loci_types:
-    count = query_type(curr, SPARQL_ENDPOINT, auth=auth)
+    count = util.query_type(curr, SPARQL_ENDPOINT, auth=auth)
     count_types.append((curr, count))
 
 print(json.dumps(count_types, indent=4, sort_keys=True))
