+++ tests/integration/test_loci_withins.py
@@ -6,42 +6,7 @@ import json
 
 load_dotenv(find_dotenv())
 
-
-def query_sfWithin(loci_thing, sparql_endpoint, auth=None):
-    '''
-        auth = expects a dict with 'user' and 'password' as keys with values,
-              e.g. { 'user': 'username', 'password': 'passwordhere' }
-    '''
-    sparql = SPARQLWrapper(sparql_endpoint)
-
-    if auth !=  None:
-        sparql.setCredentials(user=auth['user'], passwd=auth['password'])
-
-    query = '''
-        PREFIX rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
-        PREFIX geo: <http://www.opengis.net/ont/geosparql#>
-        PREFIX dct: <http://purl.org/dc/terms/>
-        select ?st ?link ?par
-        where {{
-            ?st dct:isPartOf ?lk ;
-                rdf:subject {loci_subject} ;
-                rdf:predicate geo:sfWithin ;
-                rdf:object ?par .
-        }}
-    '''.format(loci_subject=loci_thing)
-    #print(query)
-    sparql.setQuery(query)
-    sparql.setReturnFormat(JSON)
-    results = sparql.query().convert()
-    #print(json.dumps(results, indent=4, sort_keys=True))
-
-    matching_list = []
-    for r in results["results"]["bindings"]:
-        if 'par' in r:
-            matching_list.append(r['par']['value'])
-
-    #do something    
-    return matching_list
+from pyloci.sparql import util
 
 
 def get_verification_data():
@@ -69,8 +34,9 @@ def loci_relations(obj):
             "user" : GRAPHDB_USER,
             "password" : GRAPHDB_PASSWORD   
         }
-
-    matches = query_sfWithin(obj['id'], SPARQL_ENDPOINT, auth=auth)
+    
+    
+    matches = util.query_sfWithin(obj['id'], SPARQL_ENDPOINT, auth=auth)
     print(matches)
     print(obj["matches"])
